# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School Management System API",
    "version": "1.0.0",
    "description": "A comprehensive MERN stack API for managing school operations including authentication, attendance, grades, quizzes, and timetables",
    "contact": {
      "name": "API Support",
      "email": "support@schoolmanagement.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/admin/create-faculty": {
      "post": {
        "tags": ["Admin"],
        "summary": "Create faculty member",
        "description": "Create a new faculty member (Admin only)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "classes", "subject", "username", "password"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "classes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["10A", "10B", "11A"]
                  },
                  "subject": {
                    "type": "string",
                    "example": "Mathematics"
                  },
                  "username": {
                    "type": "string",
                    "example": "john.faculty"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Faculty created successfully"
          },
          "400": {
            "description": "Faculty already exists"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/create-student": {
      "post": {
        "tags": ["Admin"],
        "summary": "Create student",
        "description": "Create a new student (Admin only)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "rollNumber", "class", "section", "username", "password"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Jane Smith"
                  },
                  "rollNumber": {
                    "type": "string",
                    "example": "2024001"
                  },
                  "class": {
                    "type": "string",
                    "example": "10"
                  },
                  "section": {
                    "type": "string",
                    "example": "A"
                  },
                  "username": {
                    "type": "string",
                    "example": "jane.student"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student created successfully"
          },
          "400": {
            "description": "Student already exists"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/faculties": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get all faculties",
        "description": "Retrieve all faculty members (Admin only)",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of faculties"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/students": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get all students",
        "description": "Retrieve all students (Admin only)",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of students"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/stats": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get statistics",
        "description": "Get total count of faculties and students (Admin only)",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalFaculty": {
                      "type": "integer"
                    },
                    "totalStudents": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/attendance/students/{classSection}": {
      "get": {
        "tags": ["Attendance"],
        "summary": "Get students by class section",
        "description": "Get all students in a specific class section (Faculty only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "classSection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "10A"
          }
        ],
        "responses": {
          "200": {
            "description": "List of students"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/attendance/mark": {
      "post": {
        "tags": ["Attendance"],
        "summary": "Mark attendance",
        "description": "Mark attendance for students (Faculty only)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["attendance", "classSection", "date"],
                "properties": {
                  "attendance": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "studentId": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": ["present", "absent"]
                        }
                      }
                    }
                  },
                  "classSection": {
                    "type": "string",
                    "example": "10A"
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-12-12"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attendance marked successfully"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/attendance/student/{studentId}": {
      "get": {
        "tags": ["Attendance"],
        "summary": "Get student attendance",
        "description": "Get attendance records and statistics for a student",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attendance data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "records": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AttendanceRecord"
                      }
                    },
                    "statistics": {
                      "$ref": "#/components/schemas/AttendanceStatistics"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AttendanceRecord": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "student": {
            "type": "string"
          },
          "faculty": {
            "type": "object"
          },
          "classSection": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": ["present", "absent"]
          }
        }
      },
      "AttendanceStatistics": {
        "type": "object",
        "properties": {
          "totalDays": {
            "type": "integer"
          },
          "presentDays": {
            "type": "integer"
          },
          "absentDays": {
            "type": "integer"
          },
          "attendancePercentage": {
            "type": "string"
          }
        }
      },
...
    }
  }
}